+++ backend/test/Anela.Heblo.Tests/Features/Invoices/IssuedInvoiceRepositoryTests.cs
@@ -1,6 +1,5 @@
 using Anela.Heblo.Application.Features.Invoices.Contracts;
 using Anela.Heblo.Persistence.Invoices;
-using Anela.Heblo.Application.Features.Invoices.Infrastructure;
 using Anela.Heblo.Domain.Features.Invoices;
 using Anela.Heblo.Persistence;
 using Microsoft.EntityFrameworkCore;
